+++ R/LearnerClustPAM.R
@@ -1,8 +1,6 @@
 #' @title Partitioning Around Medoids Clustering Learner
 #'
 #' @name mlr_learners_clust.pam
-#' @include LearnerClust.R
-#' @include aaa.R
 #'
 #' @description
 #' A [LearnerClust] for PAM clustering implemented in [cluster::pam()].
@@ -80,4 +78,5 @@ LearnerClustPAM = R6Class("LearnerClustPAM",
   )
 )
 
+#' @include aaa.R
 learners[["clust.pam"]] = LearnerClustPAM